+++ packages/tokens-copy/build-tokens.js
@@ -38,6 +38,49 @@ StyleDictionary.registerTransform({
   },
 });
 
+StyleDictionary.registerTransform({
+  name: "size/rem",
+  type: "value",
+  matcher: function (token) {
+    return (
+      token.attributes.category === "sizing" ||
+      token.attributes.category === "spacing"
+    );
+  },
+  transformer: function (token) {
+    // Define the base font size in pixels (change this to match your design system)
+    const baseFontSizePx = 16;
+
+    // Calculate the value in rems
+    const valueInRem = token.original.value / baseFontSizePx;
+
+    // Round the value to a reasonable number of decimal places
+    const remValue = valueInRem.toFixed(4); // Adjust the number of decimal places as needed
+
+    return `${remValue}rem`;
+  },
+});
+
+StyleDictionary.registerTransform({
+  name: "toolkit/radii/pxToRem",
+  type: "value",
+  matcher: function (token) {
+    return token.attributes.category === "radii";
+  },
+  transformer: function (token) {
+    // Define the base font size in pixels (change this to match your design system)
+    const baseFontSizePx = 16;
+
+    // Calculate the value in rems
+    const valueInRem = token.original.value / baseFontSizePx;
+
+    // Round the value to a reasonable number of decimal places
+    const remValue = valueInRem.toFixed(4); // Adjust the number of decimal places as needed
+
+    return `${remValue}rem`;
+  },
+});
+
 const sd = StyleDictionary.extend({
   source: ["core.json"],
   platforms: {
@@ -45,7 +88,7 @@ const sd = StyleDictionary.extend({
       transforms: [
         "attribute/cti",
         "ts/descriptionToComment",
-        "ts/size/px",
+        // "ts/size/px",
         "ts/opacity",
         "ts/size/lineheight",
         "ts/type/fontWeight",
@@ -57,9 +100,10 @@ const sd = StyleDictionary.extend({
         "ts/color/css/hexrgba",
         "ts/color/modifiers",
         "name/cti/kebab",
-        "toolkit/radii/px",
-        "toolkit/spacing/px",
-        "toolkit/fontSize/px",
+        // "toolkit/radii/px",
+        // "toolkit/spacing/px",
+        "toolkit/radii/pxToRem",
+        "size/rem",
       ],
       buildPath: "build/css/",
       files: [